+++ genoss/services/model_factory.py
@@ -7,7 +7,7 @@ from genoss.llm.hf_inference_endpoint.hf_inference_endpoint import (
     HuggingFaceInferenceEndpointLLM,
 )
 from genoss.llm.local.gpt4all import Gpt4AllLLM
-from genoss.llm.openai.openai_llm import OpenAILLM
+from genoss.llm.openai_llm.openai_llm import OpenAILLM
 
 OPENAI_NAME_LIST = ["gpt-4", "gpt-3.5-turbo"]
